+++ das_client/app/integration_test/test/brake_load_slip_test.dart
@@ -10,9 +10,9 @@ import 'package:app/pages/journey/journey_screen/notification/widgets/brake_load
 import 'package:app/pages/journey/journey_screen/widgets/journey_table.dart';
 import 'package:app/pages/journey/journey_screen/widgets/table/cells/route_chevron.dart';
 import 'package:app/util/time_constants.dart';
+import 'package:app/widgets/das_circle_badge.dart';
 import 'package:app/widgets/modal_sheet/das_modal_sheet.dart';
 import 'package:app/widgets/navigation_buttons.dart';
-import 'package:app/widgets/small_circle.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:formation/component.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
