+++ das_client/app/integration_test/test/service_point_modal_test.dart
@@ -11,8 +11,8 @@ import 'package:app/pages/journey/journey_screen/header/widgets/header_icon_butt
 import 'package:app/pages/journey/journey_screen/header/widgets/journey_advancement_button.dart';
 import 'package:app/pages/journey/journey_screen/widgets/communication_network_icon.dart';
 import 'package:app/util/time_constants.dart';
+import 'package:app/widgets/das_circle_badge.dart';
 import 'package:app/widgets/modal_sheet/das_modal_sheet.dart';
-import 'package:app/widgets/small_circle.dart';
 import 'package:app/widgets/speed_display.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
