+++ das_client/app/lib/pages/journey/brake_load_slip/widgets/brake_load_slip_special_restrictions.dart
@@ -3,9 +3,9 @@ import 'package:app/theme/das_colors.dart';
 import 'package:app/theme/theme_util.dart';
 import 'package:app/widgets/assets.dart';
 import 'package:app/widgets/das_badge_overlay.dart';
+import 'package:app/widgets/das_circle_badge.dart';
 import 'package:app/widgets/key_value_table.dart';
 import 'package:app/widgets/key_value_table_data_row.dart';
-import 'package:app/widgets/small_circle.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:formation/component.dart';
