+++ das_client/app/lib/pages/journey/journey_screen/header/widgets/radio_channel.dart
@@ -3,7 +3,7 @@ import 'package:app/pages/journey/journey_screen/header/view_model/radio_channel
 import 'package:app/pages/journey/journey_screen/header/widgets/radio_contact.dart';
 import 'package:app/pages/journey/journey_screen/widgets/communication_network_icon.dart';
 import 'package:app/widgets/das_badge_overlay.dart';
-import 'package:app/widgets/small_circle.dart';
+import 'package:app/widgets/das_circle_badge.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
