+++ das_client/app/lib/widgets/key_value_table_data_row.dart
@@ -1,6 +1,6 @@
 import 'package:app/theme/theme_util.dart';
 import 'package:app/widgets/das_badge_overlay.dart';
-import 'package:app/widgets/small_circle.dart';
+import 'package:app/widgets/das_circle_badge.dart';
 import 'package:flutter/material.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
