+++ das_client/app/lib/widgets/speed_display.dart
@@ -1,5 +1,5 @@
 import 'package:app/widgets/das_badge_overlay.dart';
-import 'package:app/widgets/small_circle.dart';
+import 'package:app/widgets/das_circle_badge.dart';
 import 'package:app/widgets/table/das_table_cell.dart';
 import 'package:flutter/material.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
